# Stop removing elements from the priority queue, speeds up flexible algorithms