--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 6.0 [not yet released]
-- don't allow cars or motorcycles to use ways tagged with service=emergency_access (#2484) 
+- don't allow cars or motorcycles to use ways tagged with service=emergency_access (#2484)
+- faster flexible routing, especially in conjunction with turn costs
 
 ### 5.0 [23 Mar 2022]