--- a/core/src/main/java/com/graphhopper/routing/AbstractBidirCHAlgo.java
+++ b/core/src/main/java/com/graphhopper/routing/AbstractBidirCHAlgo.java
@@ -136,7 +136,14 @@ public abstract class AbstractBidirCHAlgo extends AbstractBidirAlgo implements B
         if (pqOpenSetFrom.isEmpty()) {
             return false;
         }
-        currFrom = pqOpenSetFrom.poll();
+        while (!pqOpenSetFrom.isEmpty()) {
+            currFrom = pqOpenSetFrom.poll();
+            if (currFrom.adjNode < 0)
+                continue;
+            break;
+        }
+        if (currFrom.adjNode < 0)
+            return false;
         visitedCountFrom++;
         if (fromEntryCanBeSkipped()) {
             return true;
@@ -154,7 +161,14 @@ public abstract class AbstractBidirCHAlgo extends AbstractBidirAlgo implements B
         if (pqOpenSetTo.isEmpty()) {
             return false;
         }
-        currTo = pqOpenSetTo.poll();
+        while (!pqOpenSetTo.isEmpty()) {
+            currTo = pqOpenSetTo.poll();
+            if (currTo.adjNode < 0)
+                continue;
+            break;
+        }
+        if (currTo.adjNode < 0)
+            return false;
         visitedCountTo++;
         if (toEntryCanBeSkipped()) {
             return true;
@@ -186,9 +200,18 @@ public abstract class AbstractBidirCHAlgo extends AbstractBidirAlgo implements B
                 bestWeightMap.put(traversalId, entry);
                 prioQueue.add(entry);
             } else if (entry.getWeightOfVisitedPath() > weight) {
-                prioQueue.remove(entry);
-                updateEntry(entry, iter.getEdge(), iter.getAdjNode(), origEdgeId, weight, currEdge, reverse);
+                // flagging this entry, so it will be ignored when it is polled the next time
+                boolean best = reverse ? (entry == bestBwdEntry) : (entry == bestFwdEntry);
+                entry.adjNode = -1;
+                entry = createEntry(iter.getEdge(), iter.getAdjNode(), origEdgeId, weight, currEdge, reverse);
+                bestWeightMap.put(traversalId, entry);
                 prioQueue.add(entry);
+                if (best) {
+                    if (reverse)
+                        bestBwdEntry = entry;
+                    else
+                        bestFwdEntry = entry;
+                }
             } else
                 continue;