@boldtrn 2022-05-09
I did test this as well and can verify your numbers. This is a serious performance improvement for flexible algos 🎉