@easbar 2022-05-16
> I did test this as well and can verify your numbers. This is a serious performance improvement for flexible algos

Thanks a lot for trying this out! Note there was a bug in combination with LM and fixed start/target edges (i.e. when using curbsides or map matching), that I just fixed in #2581.